Optimize scrollviewer with load windows
Previously, each change of index using 'a', 'd', or specifying will
result in the scrollviewer updating. However, it is not optimal. An
alternate method is to load only when the selected item to display is
near the boundary of the load window.

However, now when it is near the boundary (-/+ 10 items) from load
window, it will reload. Need to fix this in later commits