Temp fix for #2, slower frame seek but correct
Due to the OpenCV bug, we are not able to retrieve the correct frame
for comparing when using cap.set(). Alternative solution (slower)
is to call cap.grab() until we reach the desired frame.

However, this method is slower and seeking to a frame earlier than its
current position will require that the video be reloaded. E.g.
cv2.VideoCapture() again.